+++ src/apis/users/users.module.ts
@@ -3,7 +3,7 @@ import { UsersService } from './users.service';
 import { UsersController } from './users.controller';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { UserEntity } from './entities/user.entity';
-import { JwtStrategy } from 'src/auth/strategies/jwt.access.strategy';
+import { JwtStrategy } from 'src/strategies/jwt.access.strategy';
 import { DiskStorageModule } from 'src/configs/multer.module';
 
 @Module({
